Add files via upload
Problem Statement:

Plot a Histogram for the below values 
{ 10, 13, 18, 22, 27, 32, 38, 40, 45, 51, 56, 57, 88, 90, 92, 94, 99 }
Bin Size to  be used is 20
Number of Bins to be used is 5

The Histogram is present in the attached excel spreadsheet "Statistics_Assignment_1.xlsx"

Discussion point regarding the solution:

Since I had plotted it in Excel I was not able to change the number of observations for the 4th and 5th Bucket.

I feel that the 4th bucket should have only 1 observation i.e. 88 and the 5th bucket should have 4 observations i.e. 90, 92, 94 and 99.
Since the upper number of each bucket should not be inclusive.